Navigation for mobile fixed, removed check route.
